requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
pillow==10.2.0
pyobjc-core==10.1
pyobjc-framework-Cocoa==10.1
pyobjc-framework-Quartz==10.1
pystray==0.19.5
requests==2.31.0
six==1.16.0
urllib3==2.1.0
